第三次正式提交 完成了修改的提交

+++ src/main/java/com/fc/service/impl/HouseServiceImpl.java
@@ -5,12 +5,8 @@ import com.fc.entity.House;
 import com.fc.entity.HouseExample;
 import com.fc.service.HouseService;
 import com.fc.vo.ResultVo;
-import com.github.pagehelper.PageHelper;
-import com.github.pagehelper.PageInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.ui.Model;
-import org.springframework.web.servlet.ModelAndView;
 
 import java.util.List;
 
@@ -39,11 +35,11 @@ public class HouseServiceImpl implements HouseService {
     public ResultVo add(House house) {
         ResultVo resultVo = new ResultVo();
         int affectedRows = houseMapper.insertSelective(house);
-        if (affectedRows > 0){
+        if (affectedRows > 0) {
             resultVo.setCode(200);
             resultVo.setMessage("添加成功！");
             resultVo.setSuccess(true);
-        }else {
+        } else {
             resultVo.setCode(-1);
             resultVo.setSuccess(false);
             resultVo.setMessage("添加失败！");
@@ -55,47 +51,52 @@ public class HouseServiceImpl implements HouseService {
     public ResultVo update(House house) {
         ResultVo resultVo = new ResultVo();
         int affectedRows = houseMapper.updateByPrimaryKeySelective(house);
-        if (affectedRows > 0){
+        if (affectedRows > 0) {
             resultVo.setCode(200);
             resultVo.setMessage("修改信息成功！");
             resultVo.setSuccess(true);
             resultVo.setData(house);
-        }else {
+        } else {
             resultVo.setCode(-1);
             resultVo.setMessage("修改信息失败！");
             resultVo.setSuccess(false);
         }
-        return null;
+        return resultVo;
     }
 
     @Override
-    public ResultVo findById(Integer id) {
-        ResultVo resultVo = new ResultVo();
-        House house = houseMapper.selectByPrimaryKey(id);
-        if (house != null && house.equals("")){
-            resultVo.setCode(200);
-            resultVo.setSuccess(true);
-            resultVo.setData(house);
-        }else {
-            resultVo.setCode(-1);
-            resultVo.setSuccess(false);
-            resultVo.setMessage("未查到此人！");
-        }
-        return resultVo;
+    public House findById(Integer id) {
+//        ResultVo resultVo = new ResultVo();
+        return houseMapper.selectByPrimaryKey(id);
+//        if (house != null && house.equals("")) {
+//            resultVo.setCode(200);
+//            resultVo.setSuccess(true);
+//            resultVo.setData(house);
+//        } else {
+//            resultVo.setCode(-1);
+//            resultVo.setSuccess(false);
+//            resultVo.setMessage("未查到此人！");
+//        }
+//        return resultVo;
     }
 
     @Override
     public ResultVo delete(Integer id) {
         ResultVo resultVo = new ResultVo();
-        int affectedRows = houseMapper.deleteByPrimaryKey(id);
-        if (affectedRows > 0){
-            resultVo.setCode(200);
-            resultVo.setSuccess(true);
-        }else {
+        if (id == null) {
             resultVo.setCode(-1);
             resultVo.setSuccess(false);
             resultVo.setMessage("删除失败！");
+        } else {
+            int affectedRows = houseMapper.deleteByPrimaryKey(id);
+            if (affectedRows > 0) {
+                resultVo.setCode(200);
+                resultVo.setSuccess(true);
+                resultVo.setMessage("删除成功！");
+            } else {
+                System.out.println("Fail");
+            }
         }
-        return resultVo;
+            return resultVo;
     }
 }
